# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoșMic API",
    "version": "1.0.0",
    "description": "API public de citire pentru catalogul CoșMic.\n\nFiecare preț poartă data la care a fost observat. Rândurile reținute de o verificare nu sunt niciodată returnate. „Nu avem o comparație pentru acest produs” este un răspuns de sine stătător, nu o listă goală — 89% dintre produsele cu preț au un singur magazin.\n\nGENERAT din `src/lib/api/routes.ts`. Nu edita direct."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/v1/basket/optimize": {
      "post": {
        "summary": "Calculează unde este mai ieftin un coș.",
        "description": "Maxim 100 de linii, identificate prin slug. Fiecare substituție vine cu motivul ei: un coș care schimbă tăcut un produs și raportează doar totalul spune ceva fals despre ce cumperi.\n\nLimită: 30 cereri / 60s per apelant identificat (300 partajate când apelantul nu poate fi identificat).\nCache: no-store.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Parametru lipsă sau invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "bad_request",
                            "not_found",
                            "rate_limited",
                            "server_error"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nu există",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "bad_request",
                            "not_found",
                            "rate_limited",
                            "server_error"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Prea multe cereri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "bad_request",
                            "not_found",
                            "rate_limited",
                            "server_error"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}